refactor: move the AuthApi into new datasource module
Move the Auth module to a new datasource and remove the old server module. Later on the account
details can be stored in the auth module instead of within the Kitsu module
