+++ app/src/androidTest/java/com/chesire/nekome/features/login/LoginFlowTests.kt
@@ -1,8 +1,8 @@
 package com.chesire.nekome.features.login
 
 import com.chesire.nekome.UITest
-import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.auth.remote.AuthApi
 import com.chesire.nekome.datasource.series.remote.SeriesApi
 import com.chesire.nekome.datasource.user.remote.UserApi
 import com.chesire.nekome.helpers.creation.createSeriesDomain
