+++ app/src/main/java/com/chesire/nekome/services/RefreshAuthWorker.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.auth.remote.AuthApi
 import com.chesire.nekome.datasource.user.UserRepository
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
