+++ features/login/build.gradle
@@ -32,9 +32,9 @@ android {
 
 dependencies {
     implementation project(":libraries:core")
+    implementation project(":libraries:datasource:auth")
     implementation project(":libraries:datasource:series")
     implementation project(":libraries:datasource:user")
-    implementation project(":libraries:server:auth")
 
     implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
