+++ features/login/src/main/java/com/chesire/nekome/app/login/details/DetailsViewModel.kt
@@ -3,8 +3,8 @@ package com.chesire.nekome.app.login.details
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.auth.remote.AuthApi
 import com.chesire.nekome.datasource.user.UserRepository
 import com.hadilq.liveevent.LiveEvent
 import dagger.hilt.android.lifecycle.HiltViewModel
