+++ features/login/src/test/java/com/chesire/nekome/app/login/details/DetailsViewModelTests.kt
@@ -2,8 +2,8 @@ package com.chesire.nekome.app.login.details
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.auth.remote.AuthApi
 import com.chesire.nekome.datasource.user.UserRepository
 import com.chesire.nekome.testing.CoroutinesMainDispatcherRule
 import io.mockk.Runs
