+++ libraries/datasource/auth/.gitignore
