+++ libraries/kitsu/build.gradle
@@ -31,8 +31,8 @@ android {
 
 dependencies {
     implementation project(":libraries:core")
+    implementation project(":libraries:datasource:auth")
     implementation project(":libraries:encryption")
-    implementation project(":libraries:server:auth")
 
     implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "androidx.core:core-ktx:$corektx_version"
