--- libraries/server/auth/src/main/java/com/chesire/nekome/auth/api/AuthDomain.kt
@@ -1,7 +1,0 @@
-package com.chesire.nekome.auth.api
-
-// Currently not needed - will be used in a refactor #387
-/**
- * Provides a domain for the Auth module.
- */
-class AuthDomain
